removed history responsibility from Calculator and moved it to a new class History
